+++ Package.swift
@@ -19,8 +19,8 @@ let package: Package = .init(
     targets: [
         .binaryTarget(
             name: "VCore",
-            url: "https://github.com/VakhoKontridze/VCore/releases/download/1.2.0/VCore.xcframework.zip",
-            checksum: "60949bdc10b8877a17b3be53cc74f0e228b628145d4e052699d26fc7cfb39d02"
+            url: "https://github.com/VakhoKontridze/VCore/releases/download/1.3.0/VCore.xcframework.zip",
+            checksum: "999bcf23945b68cfb4b2df47e28cdccdac70c8c5a59d7d0f5d5d47302c3b0880"
         ),
     ]
 )