+++ VCore/Helpers/Deprecations.swift
@@ -5,4 +5,4 @@
 //  Created by Vakhtang Kontridze on 10/9/21.
 //
 
-import Foundation
+import UIKit